Refactored mess.

+++ tests/rackspace/models/block_storage/snapshot_tests.rb
@@ -1,20 +1,25 @@
 Shindo.tests('Fog::Rackspace::BlockStorage | snapshot', ['rackspace']) do
-
-  pending if Fog.mocking?
-
   service = Fog::Rackspace::BlockStorage.new
-  volume = service.volumes.create({
-    :display_name => "fog_#{Time.now.to_i.to_s}",
-    :size => 100
-  })
 
-  volume.wait_for { ready? }
+  begin
+    volume = service.volumes.create({
+      :display_name => "fog_#{Time.now.to_i.to_s}",
+      :size => 100
+    })
 
-  options = { :display_name => "fog_#{Time.now.to_i.to_s}", :volume_id => volume.id }
-  model_tests(service.snapshots, options, false) do
-    @instance.wait_for { ready? }
+    volume.wait_for { ready? }
+
+    options = { :display_name => "fog_#{Time.now.to_i.to_s}", :volume_id => volume.id }
+    model_tests(service.snapshots, options, true) do
+      @instance.wait_for { ready? }
+
+      tests('double save').raises(Fog::Rackspace::BlockStorage::IdentifierTaken) do
+        @instance.save
+      end
+    end
+
+    volume.wait_for { snapshots.empty? }
+  ensure
+    volume.destroy if volume
   end
-
-  volume.wait_for { snapshots.empty? }
-  volume.destroy
 end
