+++ tests/rackspace/models/block_storage/volume_tests.rb
@@ -1,12 +1,14 @@
 Shindo.tests('Fog::Rackspace::BlockStorage | volume', ['rackspace']) do
 
-  pending if Fog.mocking?
-
   service = Fog::Rackspace::BlockStorage.new
   options = { :display_name => "fog_#{Time.now.to_i.to_s}", :size => 100 }
 
-  model_tests(service.volumes, options, false) do
-    @instance.wait_for { ready? }
+  model_tests(service.volumes, options, true) do
+    @instance.wait_for{ ready? }
+
+    tests('double save').raises(Fog::Rackspace::BlockStorage::IdentifierTaken) do
+      @instance.save
+    end
 
     tests('#attached?').succeeds do
       @instance.state = 'in-use'
@@ -14,12 +16,14 @@ Shindo.tests('Fog::Rackspace::BlockStorage | volume', ['rackspace']) do
     end
 
     tests('#snapshots').succeeds do
-      snapshot = service.snapshots.create({ :volume_id => @instance.id })
-      snapshot.wait_for { ready? }
+      begin
+        snapshot = @instance.create_snapshot
+        snapshot.wait_for { ready? }
 
-      returns(true) { @instance.snapshots.first.id == snapshot.id }
-
-      snapshot.destroy
+        returns(true) { @instance.snapshots.first.id == snapshot.id }
+      ensure
+        snapshot.destroy if snapshot
+      end
     end
 
     @instance.wait_for { snapshots.empty? }
